pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glpi-isms-toolkit"
version = "1.0.0"
description = "GLPI ITSM bootstrap toolkit with ISO 27001 ISMS integration, configuration-as-code, and executive report generation."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Erdem Özyurt", email = "erdem@siberkale.com" },
]
keywords = ["glpi", "itsm", "itam", "iso27001", "isms", "itil", "helpdesk", "asset-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
    "Topic :: Office/Business",
]
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "reportlab>=4.0",
    "python-pptx>=0.6.21",
    "requests>=2.28",
    "rich>=13.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
glpi-toolkit = "glpi_toolkit.cli:app"

[project.urls]
Homepage = "https://github.com/erdemozyurt/glpi-isms-toolkit"
Repository = "https://github.com/erdemozyurt/glpi-isms-toolkit"
Issues = "https://github.com/erdemozyurt/glpi-isms-toolkit/issues"
Documentation = "https://github.com/erdemozyurt/glpi-isms-toolkit/tree/main/docs"

[tool.hatch.build.targets.wheel]
packages = ["src/glpi_toolkit"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
